# api_gateway/OpenAPI.yml
openapi: 3.0.3
info:
  title: StudyFlow API
  version: 1.0.0
components:
  schemas:
    UserRole:
      type: string
      enum:
        - tutor
        - student
...
    TutorStudentStatus:
      type: string
      enum:
        - invited
        - active
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        firstName:
          type: string
        lastName:
          type: string
    TutorProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        paymentInfo:
          type: string
        lessonPriceRub:
          type: integer
        lessonConnectionLink:
          type: string
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    TutorStudent:
      type: object
      properties:
        id:
          type: string
        tutorId:
          type: string
        studentId:
          type: string
        lessonPriceRub:
          type: integer
        lessonConnectionLink:
          type: string
        status:
          $ref: '#/components/schemas/TutorStudentStatus'
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        message:
          type: string
...
paths:
...
  /users/users/{id}:
    get:
      summary: Get public user profile
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User public info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update user info
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                timezone:
                  type: string
      responses:
        '200':
          description: User updated
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-profiles/{user_id}:
    get:
      summary: Get tutor profile by user_id
      operationId: getTutorProfile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tutor profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorProfile'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update tutor profile
      operationId: updateTutorProfile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paymentInfo:
                  type: string
                lessonPriceRub:
                  type: integer
                lessonConnectionLink:
                  type: string
      responses:
        '200':
          description: Tutor profile updated
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students/by-tutor/{tutor_id}:
    get:
      summary: List tutor-student relationships by tutor
      operationId: listTutorStudents
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TutorStudent'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students/by-student/{student_id}:
    get:
      summary: List tutor-student relationships by student
      operationId: listTutorsForStudent
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TutorStudent'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students/{tutor_id}/{student_id}:
    get:
      summary: Get tutor-student relationship
      operationId: getTutorStudent
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Relationship found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorStudent'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update tutor-student relationship
      operationId: updateTutorStudent
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lessonPriceRub:
                  type: integer
                lessonConnectionLink:
                  type: string
                status:
                  $ref: '#/components/schemas/TutorStudentStatus'
      responses:
        '200':
          description: Relationship updated
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete tutor-student relationship
      operationId: deleteTutorStudent
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Relationship deleted
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students/{tutor_id}/accept:
    post:
      summary: Accept tutor invitation
      operationId: acceptInvitationFromTutor
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation accepted
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tutor or relationship not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students:
    post:
      summary: Create tutor-student relationship
      operationId: createTutorStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tutorId:
                  type: string
                studentId:
                  type: string
                lessonPriceRub:
                  type: integer
                lessonConnectionLink:
                  type: string
              required:
                - tutor_id
                - student_id
      responses:
        '200':
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorStudent'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Relationship already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


  # files
...